- split SliderAndIndicatorBlocBuilder to new section
