+++ lib/features/home/ui/screens/home_screen.dart
@@ -1,12 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:netflix_clone/core/helpers/spacing.dart';
 import 'package:netflix_clone/core/theming/app_colors.dart';
-import 'package:netflix_clone/features/home/logic/cubit/home_cubit.dart';
-import 'package:netflix_clone/features/home/logic/cubit/home_state.dart';
-import 'package:netflix_clone/features/home/ui/widgets/slider_and_indicator_section.dart';
-
+import 'package:netflix_clone/features/home/ui/widgets/slider_section/slider_and_indicator_bloc_builder.dart';
 import '../../../../core/helpers/constants.dart';
 
 class HomeScreen extends StatelessWidget {
@@ -39,29 +35,12 @@ class HomeScreen extends StatelessWidget {
             horizontalSpace(15)
           ],
         ),
-        body: SingleChildScrollView(
+        body: const SingleChildScrollView(
           child: Column(
             children: [
-              BlocBuilder<HomeCubit, HomeState>(
-                buildWhen: (previous, current) =>
-                    current is PopularMoviesLoading ||
-                    current is PopularMoviesSuccess ||
-                    current is PopularMoviesFailure,
-                builder: (context, state) {
-                  if (state is PopularMoviesLoading) {
-                    return const Center(
-                      child: CircularProgressIndicator(),
-                    );
-                  } else if (state is PopularMoviesSuccess) {
-                    return SliderAndIndicatorSection(list: state.movies.sublist(0, 5));
-                    
-                  } else if (state is PopularMoviesFailure) {
-                    return Text(state.error);
-                  } else {
-                    return const SizedBox.shrink();
-                  }
-                },
-              )
+
+              SliderAndIndicatorBlocBuilder()
+
             ],
           ),
         ));
